Fix travis build
Fixes #1 and #51

I'm not happy with the fact I had to remove the stringer call from the
Makefile. I don't understand why that doesn't work yet but it was just
getting on the way. I knew that getting the build green would take more
work.

I think this is a good start but it'll need some refinements in a few
weeks.